+++ src/test/java/com/epam/reportportal/junit/utils/TestUtils.java
@@ -172,6 +172,7 @@ public class TestUtils {
 		result.setLaunchName("My-test-launch" + generateUniqueId());
 		result.setProjectName("test-project");
 		result.setEnable(true);
+		result.setCallbackReportingEnabled(true);
 		return result;
 	}